requirements.txt
Flask
flask
flask-admin
flask-security sqlalchemy
flask-login
flask-script
flask-migrate
flask_table
Flask-Bootstrap4
Jinja2
MarkupSafe
SQLAlchemy
sqlalchemy
Flask-SQLAlchemy
Flask-MongoEngine
Flask-Flywheel
Werkzeug
itsdangerous
jsonschema
nose
psycopg2
mistune